ref(frontend): Split SSR server into focused runtime modules
Extract request metadata, logging, static asset serving, request
validation, internal cache invalidation, backend proxying, and document
rendering into dedicated server modules.

Keep index.mjs as the thin HTTP bootstrap and dispatcher so the SSR
runtime is easier to test and change without mixing unrelated concerns
in a single file.
